size

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte", (unsigned long)sizeof(char));
-	printf("Size of an int: %lu byte", (unsigned long)sizeof(int));
-	printf("Size of a long int: %lu byte", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu byte", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %lu byte", (unsigned long)sizeof(float));
+	printf("Size of a char: %lu byte", sizeof(char));
+	printf("Size of an int: %lu byte", sizeof(int));
+	printf("Size of a long int: %lu byte", sizeof(long int));
+	printf("Size of a long long int: %lu byte", sizeof(long long int));
+	printf("Size of a float: %lu byte", sizeof(float));
 		return (0);
 }
